+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.0
 - New settings to launch device shell/logs either in integrated or external terminal
 - Check minimum leaf version (1.6)
+- Automatically import Legato snippets from leaf packages
 
 ## 0.4.0
 - New build task to trigger build+install in one step